+++ aws-distro-opentelemetry-node-autoinstrumentation/src/instrumentation/instrumentation-langchain/instrumentation.ts
@@ -19,7 +19,7 @@ import {
 
 export const INSTRUMENTATION_NAME = '@aws/aws-distro-opentelemetry-instrumentation-langchain';
 export const INSTRUMENTATION_SHORT_NAME = 'aws_langchain';
-const SUPPORTED_VERSIONS = ['>=0.3.73 <2.0.0'];
+const SUPPORTED_VERSIONS = ['>=1.0.0 <2.0.0'];
 
 export class LangChainInstrumentation extends InstrumentationBase<LangChainInstrumentationConfig> {
   _patchedCallbackManagers: Set<any> = new Set();